Add CLI option, and output JSON
